fix(el3-runtime): resolve essential-type mismatch
This resolves MISRA C:2012 Rule 10.7 violation
where a composite expression involved operands of differing
essential types, causing unintended implicit conversions.

The fix ensures all operands in the expression have matching
essential types by introducing explicit casts,
preventing unsafe or inconsistent arithmetic operations.

+++ plat/xilinx/zynqmp/sip_svc_setup.c
@@ -121,6 +121,6 @@ DECLARE_RT_SVC(
 		sip_svc,
 		OEN_SIP_START,
 		OEN_SIP_END,
-		(uint8_t)SMC_TYPE_FAST,
+		SMC_TYPE_FAST,
 		sip_svc_setup,
 		sip_svc_smc_handler);
